+++ Cargo.toml
@@ -9,7 +9,7 @@ base64 = "0.22.1"
 clap = { version = "4.5.28", features = ["derive"] }
 colored = "2.0.0"
 csv = "1.3"
-regex = "1.11.1"
+regex = "1.11"
 reqwest = { version = "0.11.6", features = ["blocking", "json"] }
 serde = { version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.68"
